BAVL-25 pulling out classes into their own separate files.
