+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsbookavideolinkapi/service/events/ManageExternalAppointmentsService.kt
@@ -1,6 +1,7 @@
 package uk.gov.justice.digital.hmpps.hmppsbookavideolinkapi.service.events
 
 import jakarta.persistence.EntityNotFoundException
+import org.slf4j.LoggerFactory
 import org.springframework.stereotype.Service
 import uk.gov.justice.digital.hmpps.hmppsbookavideolinkapi.client.activitiesappointments.ActivitiesAppointmentsClient
 import uk.gov.justice.digital.hmpps.hmppsbookavideolinkapi.client.prisonapi.PrisonApiClient
@@ -22,8 +23,14 @@ class ManageExternalAppointmentsService(
   private val prisonApiClient: PrisonApiClient,
 ) {
 
+  companion object {
+    private val log = LoggerFactory.getLogger(this::class.java)
+  }
+
   // TODO: Assumes one person per booking, so revisit for co-defendant cases
   fun createAppointments(videoBookingId: Long) {
+    log.info("EXTERNAL APPOINTMENTS: creating appointments for video booking ID $videoBookingId")
+
     val videoBooking = videoBookingRepository.findById(videoBookingId)
       .orElseThrow { EntityNotFoundException("Video booking with ID $videoBookingId not found") }
 
@@ -33,8 +40,12 @@ class ManageExternalAppointmentsService(
 
     if (activitiesAppointmentsClient.isAppointmentsRolledOutAt(prisonCode)) {
       activitiesAppointmentsClient.createAppointment()
+
+      log.info("EXTERNAL APPOINTMENTS: created appointments for video booking ID $videoBookingId in activities and appointments")
     } else {
       prisonApiClient.createAppointment()
+
+      log.info("EXTERNAL APPOINTMENTS: created appointments for video booking ID $videoBookingId in prison api")
     }
   }
 }
